refactor(core): remember-tool.ts 분해 — 1694줄/complexity 111 핸들러 모듈 분리 (#582)
- remember-tool.ts 1694줄 → 251줄 (85% 감소)
- cyclomatic complexity 111 → handle()당 분기 제거
- 9개 핸들러 모듈로 분리:
  - remember-tool-schema.ts: Zod 스키마/타입
  - remember-tool-host.ts: RememberToolHost 인터페이스
  - remember-tool-types.ts: 공유 DB 타입
  - remember-tool-core.ts: Core Memory 핸들러
  - remember-tool-vault.ts: Knowledge Vault 핸들러
  - remember-tool-memory-item.ts: working/episodic/semantic/procedural 핸들러
  - remember-tool-reflection.ts: reflection_notes 처리
  - remember-tool-db-helpers.ts: DB 쿼리 헬퍼
  - remember-tool-augmentation.ts: 백그라운드 증강 파이프라인
- recall-tool 패턴(RecallToolHost) 동일 적용
- 테스트 호환 wrapper 유지 (getMemoryById, getExistingMemoriesForRelationExtraction)
- 전체 88 remember 테스트 통과
